refactor image pair based codes
