+++ autode/bracket/imagepair.py
@@ -1,12 +1,11 @@
+from abc import ABC
 from typing import Optional, List, Tuple
 import numpy as np
 
 from autode.values import Distance, PotentialEnergy, Gradient
 from autode.opt.coordinates import OptCoordinates, CartesianCoordinates
 from autode.opt.optimisers.base import _OptimiserHistory
-from autode.utils import work_in_tmp_dir, ProcessPool
-from autode.log import logger
-from autode.config import Config
+from autode.utils import work_in_tmp_dir
 
 import autode.species.species
 import autode.wrappers.methods
@@ -56,21 +55,21 @@ def _calculate_hessian_for_species(
     return species.hessian.to("ha/ang^2")
 
 
-class DistanceConstrainedImagePair:
+class BaseImagePair(ABC):
     """
-    A pair of images (reactant and product) of the same species,
-    constrained by Euclidean distance between their coordinates
-    (d = d_i), where d_i is a constant. The constraint is enforced
-    by Lagrangian multipliers
+    Base class for a pair of images (e.g., reactant and product) of
+    the same species. The images are called 'left' and 'right' to
+    distinguish them, but there is no requirement for one to be
+    reactant or product.
     """
+
     def __init__(
         self,
         left_image: autode.species.species.Species,
         right_image: autode.species.species.Species,
     ):
         """
-        Initialize the constrained image pair. Does not initialize
-        methods!
+        Initialize the image pair, does not set methods/n_cores!
 
         Args:
             left_image: One molecule of the pair
@@ -80,21 +79,52 @@ class DistanceConstrainedImagePair:
         assert isinstance(right_image, autode.species.species.Species)
         self._left_image = left_image.new_species(name="left_image")
         self._right_image = right_image.new_species(name="right_image")
-        # todo sanity check that two images have same charge etc. and they are not
-        # too close
+        self._sanity_check()
 
+        # separate methods for engrad and hessian calc
         self._engrad_method = None
         self._hess_method = None
         self._n_cores = None
 
-        # C_d = (d - d_i)
-        self._d_i = None
-        self._lambda_dist = 0  # dist. constraint lagrange multiplier
-
         self._left_history = _OptimiserHistory()
         self._right_history = _OptimiserHistory()
-        self.left_coord = self._left_image.coordinates.to('ang').flatten()
-        self.right_coord = self._right_image.coordinates.to('ang').flatten()
+        self.left_coord = self._left_image.coordinates.to("ang").flatten()
+        self.right_coord = self._right_image.coordinates.to("ang").flatten()
+
+    def _sanity_check(self) -> None:
+        """
+        Check if the two supplied images have the same
+        charge, multiplicity and the same atoms in the same order
+        """
+
+        if self._left_image.n_atoms != self._right_image.n_atoms:
+            raise ValueError(
+                "The initial_species and final_species must "
+                "have the same number of atoms!"
+            )
+
+        if (
+            self._left_image.charge != self._right_image.charge
+            or self._left_image.mult != self._right_image.mult
+        ):
+            raise ValueError(
+                "Charge/multiplicity of initial_species and "
+                "final_species supplied are not the same"
+            )
+
+        for idx in range(len(self._left_image.atoms)):
+            if (
+                self._left_image.atoms[idx].label
+                != self._right_image.atoms[idx].label
+            ):
+                raise ValueError(
+                    "The order of atoms in initial_species "
+                    "and final_species must be the same. The "
+                    f"atom at position {idx} is different in"
+                    "the two species"
+                )
+
+        return None
 
     def set_method_and_n_cores(
         self,
@@ -118,6 +148,71 @@ class DistanceConstrainedImagePair:
         self._hess_method = hess_method
         self._n_cores = int(n_cores)
 
+    @property
+    def n_atoms(self):
+        """Number of atoms"""
+        return self._left_image.n_atoms
+
+    @property
+    def left_coord(self) -> Optional[CartesianCoordinates]:
+        """The coordinates of the left image"""
+        if len(self._left_history) == 0:
+            return None
+        return self._left_history[-1]
+
+    @left_coord.setter
+    def left_coord(self, value):
+        """
+        Sets the coordinates of the left image, also updates
+        the coordinates of the species
+
+        Args:
+            value (np.ndarray|CartesianCoordinates): flat or (n_atoms, 3) array
+                                                     of coordinates
+        """
+        if value is None:
+            return
+        elif isinstance(value, CartesianCoordinates):
+            self._left_history.append(value.copy())
+        elif (
+            isinstance(value, np.ndarray)
+            and value.flatten().shape[0] == 3 * self.n_atoms
+        ):
+            self._right_history.append(CartesianCoordinates(value))
+        else:
+            raise ValueError
+        self._left_image.coordinates = self._left_history[-1]
+
+    @property
+    def right_coord(self) -> Optional[CartesianCoordinates]:
+        """The coordinates of the right image"""
+        if len(self._right_history) == 0:
+            return None
+        return self._right_history[-1]
+
+    @right_coord.setter
+    def right_coord(self, value):
+        """
+        Sets the coordinates of the left image, also updates
+        the coordinates of the species
+
+        Args:
+            value (np.ndarray|CartesianCoordinates): flat or (n_atoms, 3) array
+                                                     of coordinates
+        """
+        if value is None:
+            return
+        elif isinstance(value, CartesianCoordinates):
+            self._right_history.append(value.copy())
+        elif (
+            isinstance(value, np.ndarray)
+            and value.flatten().shape[0] == 3 * self.n_atoms
+        ):
+            self._right_history.append(CartesianCoordinates(value))
+        else:
+            raise ValueError
+        self._right_image.coordinates = self._right_history[-1]
+
     def update_one_side_molecular_engrad(self, side: str):
         assert self._engrad_method is not None
         assert self._n_cores is not None
@@ -135,9 +230,59 @@ class DistanceConstrainedImagePair:
             method=self._engrad_method,
             n_cores=self._n_cores,
         )
+        # update both species and coord
         img.energy = en
         img.gradient = grad.copy()
+        coord.e = en
         coord.update_g_from_cart_g(grad)
+        return None
+
+    def update_one_side_molecular_hessian(self, side: str) -> None:
+        assert self._hess_method is not None
+        assert self._n_cores is not None
+        if side == "left":
+            img = self._left_image
+            coord = self.left_coord
+        elif side == "right":
+            img = self._right_image
+            coord = self.right_coord
+        else:
+            raise Exception
+
+        hess = _calculate_hessian_for_species(
+            species=img.copy(), method=self._hess_method, n_cores=self._n_cores
+        )
+        img.hessian = hess
+        coord.update_h_from_cart_h(hess)
+        return None
+
+
+class DistanceConstrainedImagePair(BaseImagePair):
+    """
+    A pair of images (reactant and product) of the same species,
+    constrained by Euclidean distance between their coordinates
+    (d = d_i), where d_i is a constant. The constraint is enforced
+    by Lagrangian multipliers
+    """
+
+    def __init__(
+        self,
+        left_image: autode.species.species.Species,
+        right_image: autode.species.species.Species,
+    ):
+        """
+        Initialize the constrained image pair. Does not initialize
+        methods!
+
+        Args:
+            left_image: One molecule of the pair
+            right_image: Another molecule of the pair
+        """
+        super().__init__(left_image, right_image)
+
+        # C_d = (d - d_i)
+        self._d_i = None
+        self._lambda_dist = 0  # dist. constraint lagrange multiplier
 
     def get_one_sided_jacobian_of_constraints(self, side: str) -> np.ndarray:
         """
@@ -203,25 +348,6 @@ class DistanceConstrainedImagePair:
         # grad(L) = [[grad_con, -C_d]]
         grad_L = np.vstack((grad_con, constr_func_col))
         return grad_L.flatten()
-
-    def update_one_side_molecular_hessian(self, side: str) -> None:
-        assert self._hess_method is not None
-        assert self._n_cores is not None
-        if side == "left":
-            img = self._left_image
-            coord = self.left_coord
-        elif side == "right":
-            img = self._right_image
-            coord = self.right_coord
-        else:
-            raise Exception
-
-        hess = _calculate_hessian_for_species(
-            species=img.copy(), method=self._hess_method, n_cores=self._n_cores
-        )
-        img.hessian = hess
-        coord.update_h_from_cart_h(hess)
-        return None
 
     def get_one_sided_hessians_of_constraints(
         self, side: str
@@ -308,49 +434,8 @@ class DistanceConstrainedImagePair:
         return lagrange_hess
 
     @property
-    def n_atoms(self):
-        """Number of atoms"""
-        return self._left_image.n_atoms
-
-    @property
     def n_constraints(self) -> int:
-        return 2 if self._is_e_constr else 1
-
-    @property
-    def left_coord(self) -> Optional[CartesianCoordinates]:
-        if len(self._left_history) == 0:
-            return None
-        return self._left_history[-1]
-
-    @left_coord.setter
-    def left_coord(self, value):
-        if value is None:
-            return
-        elif isinstance(value, CartesianCoordinates):
-            self._left_history.append(value.copy())
-        elif isinstance(value, np.ndarray) and value.shape == 3 * self.n_atoms:
-            self._right_history.append(CartesianCoordinates(value))
-        else:
-            raise ValueError
-        self._left_image.coordinates = self._left_history[-1]
-
-    @property
-    def right_coord(self) -> Optional[CartesianCoordinates]:
-        if len(self._right_history) == 0:
-            return None
-        return self._right_history[-1]
-
-    @right_coord.setter
-    def right_coord(self, value):
-        if value is None:
-            return
-        elif isinstance(value, CartesianCoordinates):
-            self._right_history.append(value.copy())
-        elif isinstance(value, np.ndarray) and value.shape == 3 * self.n_atoms:
-            self._right_history.append(CartesianCoordinates(value))
-        else:
-            raise ValueError
-        self._right_image.coordinates = self._right_history[-1]
+        return 1
 
     @property
     def target_dist(self) -> Distance:
@@ -405,170 +490,3 @@ class DistanceConstrainedImagePair:
         """
         lambda_val = float(num)
         self._lambda_dist = lambda_val
-
-
-_optional_method = Optional[autode.wrappers.methods.Method]
-
-
-class DHS:
-    def __init__(
-        self,
-        initial_species: autode.species.Species,
-        final_species: autode.species.Species,
-        maxiter: int = 10,
-        reduction_factor: float = 0.05,
-        dist_tol: float = 0.05,
-    ):
-        self.imgpair = DistanceConstrainedImagePair(
-            initial_species, final_species
-        )
-        self._reduction_fac = reduction_factor
-
-        if int(maxiter) <= 0:
-            raise ValueError(
-                "An optimiser must be able to run at least one "
-                f"step, but tried to set maxiter = {maxiter}"
-            )
-
-        self._maxiter = int(maxiter)
-        self._dist_tol = Distance(dist_tol, units="ang")
-        self._engrad_method = None
-        self._hess_method = None
-        self._n_cores = None
-
-        self._grad = None
-        self._hess = None
-
-        # todo have a history where store the optimised points after each macroiter
-
-    @property
-    def macroiter_converged(self):
-        if self.imgpair.euclid_dist <= self._dist_tol:
-            return True
-        else:
-            return False
-
-    @property
-    def microiter_converged(self):
-        # gtol and dist_tol
-        pass
-
-    @property
-    def _exceeded_maximum_iteration(self) -> bool:
-        return True if self.imgpair.total_iters > self._maxiter else False
-
-    def calculate(
-        self,
-        engrad_method: _optional_method = None,
-        hess_method: _optional_method = None,
-        n_cores: Optional[int] = None,
-    ):
-        from autode.methods import (
-            method_or_default_hmethod,
-            method_or_default_lmethod,
-        )
-
-        engrad_method = method_or_default_hmethod(engrad_method)
-        hess_method = method_or_default_lmethod(hess_method)
-        if n_cores is None:
-            n_cores = Config.n_cores
-        self.imgpair.set_method_and_n_cores(
-            engrad_method=engrad_method,
-            hess_method=hess_method,
-            n_cores=n_cores,
-        )
-
-        self._initialise_run()
-
-        # in each macroiter, the distance criteria is reduced by factor
-        macroiter_num = 0
-        maxiter_reached = False
-        while not self.macroiter_converged:
-            self.imgpair.target_dist = self.imgpair.euclid_dist * (
-                1 - self._reduction_fac
-            )
-            macroiter_num += 1
-            while not self.microiter_converged:
-                if self.imgpair.left_coord.e < self.imgpair.right_coord.e:
-                    side = "left"
-                else:
-                    side = "right"
-                self._take_one_sided_step(side)
-                self._update_one_side_mol_engrad_hess(side)
-                # in gradient update step, update hessian as well
-                # no need to update hessian in prfo step
-                if self._exceeded_maximum_iteration:
-                    # error message
-                    maxiter_reached = True
-                    break
-            if maxiter_reached:
-                break
-
-        pass
-
-    def _take_one_sided_step(self, side: str):
-        grad = self.imgpair.get_one_sided_lagrangian_gradient(side)
-        hess = self.imgpair.get_one_sided_lagrangian_hessian(side)
-        self._prfo_step(side, grad, hess)
-
-    def _update_one_side_mol_engrad_hess(self, side: str):
-        self.imgpair.update_one_side_molecular_engrad(side)
-        # todo interpolate hessian by Bofill
-        #self.imgpair.update_one_side_molecular_hessian(side)
-
-    def _initialise_run(self):
-        # todo this func is empty
-        # todo is it really necessary to parallelise in DHS?
-        self.imgpair.update_one_side_molecular_engrad('left')
-        self.imgpair.update_one_side_molecular_hessian('left')
-        self.imgpair.update_one_side_molecular_engrad('right')
-        self.imgpair.update_one_side_molecular_hessian('right')
-        pass
-
-    def _prfo_step(self, side: str, grad: np.ndarray, hess:np.ndarray):
-        b, u = np.linalg.eigh(hess)
-        f = u.T.dot(grad)
-
-        delta_s = np.zeros(shape=(self.imgpair.n_atoms,))
-
-        # partition the RFO
-        # The constraint is the last item
-        constr_components = u[-1, :]
-        # highest component of constraint must be the required mode
-        constr_mode = np.argmax(constr_components)
-        u_max = u[:, constr_mode]
-        b_max = b[constr_mode]
-        f_max = f[constr_mode]
-
-        # only one constraint - distance
-        aug_h_max = np.zeros(shape=(2, 2))
-        aug_h_max[0, 0] = b_max
-        aug_h_max[1, 0] = f_max
-        aug_h_max[0, 1] = f_max
-        lambda_p = np.linalg.eigvalsh(aug_h_max)[-1]
-
-        delta_s -= f_max * u_max / (b_max - lambda_p)
-
-        u_min = np.delete(u, constr_mode, axis=1)
-        b_min = np.delete(b, constr_mode)
-        f_min = np.delete(f, constr_mode)
-        # todo deal with rot./trans.?
-
-        # n_atoms non-constraint modes
-        m = self.imgpair.n_atoms
-        aug_h_min = np.zeros(shape=(m + 1, m + 1))
-        aug_h_min[:m, :m] = np.diag(b_min)
-        aug_h_min[-1, :m] = f_min
-        aug_h_min[:m, -1] = f_min
-        eigvals_min = np.linalg.eigvalsh(aug_h_min)
-        min_mode = np.where(np.abs(eigvals_min) > 1e-15)[0][0]
-        lambda_n = eigvals_min[min_mode]
-
-        for i in range(m):
-            delta_s -= f_min[i] * u_min[:, i] / (b_min[i] - lambda_n)
-
-        self.take_step_within_trust_radius(side, delta_s)
-
-    def take_step_within_trust_radius(self, side, delta_s):
-        # set the coords to one side
-        pass
